Create useActivityList hook, move types to own folder and delete unused useCampaignActivites hook.
