+++ src/features/campaigns/hooks/useActivityOverview.ts
@@ -3,62 +3,13 @@ import useCallAssignmentActivities from './useCallAssignmentActivities';
 import useEventsFromDateRange from 'features/events/hooks/useEventsFromDateRange';
 import useSurveyActivities from './useSurveyActivities';
 import useTaskActivities from './useTaskActivities';
+import { ACTIVITIES, ActivityOverview, CampaignActivity } from '../types';
 import {
   ErrorFuture,
   IFuture,
   LoadingFuture,
   ResolvedFuture,
 } from 'core/caching/futures';
-import {
-  ZetkinCallAssignment,
-  ZetkinEvent,
-  ZetkinSurvey,
-  ZetkinTask,
-} from 'utils/types/zetkin';
-
-export enum ACTIVITIES {
-  CALL_ASSIGNMENT = 'callAssignment',
-  EVENT = 'event',
-  SURVEY = 'survey',
-  TASK = 'task',
-}
-
-type CampaignActivityBase = {
-  visibleFrom: Date | null;
-  visibleUntil: Date | null;
-};
-
-export type CallAssignmentActivity = CampaignActivityBase & {
-  data: ZetkinCallAssignment;
-  kind: ACTIVITIES.CALL_ASSIGNMENT;
-};
-
-export type SurveyActivity = CampaignActivityBase & {
-  data: ZetkinSurvey;
-  kind: ACTIVITIES.SURVEY;
-};
-
-export type TaskActivity = CampaignActivityBase & {
-  data: ZetkinTask;
-  kind: ACTIVITIES.TASK;
-};
-
-export type EventActivity = CampaignActivityBase & {
-  data: ZetkinEvent;
-  kind: ACTIVITIES.EVENT;
-};
-
-export type CampaignActivity =
-  | CallAssignmentActivity
-  | EventActivity
-  | SurveyActivity
-  | TaskActivity;
-
-export type ActivityOverview = {
-  alsoThisWeek: CampaignActivity[];
-  today: CampaignActivity[];
-  tomorrow: CampaignActivity[];
-};
 
 export default function useActivitiyOverview(
   orgId: number,
