+++ src/features/campaigns/types.ts
@@ -1,0 +1,50 @@
+import {
+  ZetkinCallAssignment,
+  ZetkinEvent,
+  ZetkinSurvey,
+  ZetkinTask,
+} from 'utils/types/zetkin';
+
+export enum ACTIVITIES {
+  CALL_ASSIGNMENT = 'callAssignment',
+  EVENT = 'event',
+  SURVEY = 'survey',
+  TASK = 'task',
+}
+
+type CampaignActivityBase = {
+  visibleFrom: Date | null;
+  visibleUntil: Date | null;
+};
+
+export type CallAssignmentActivity = CampaignActivityBase & {
+  data: ZetkinCallAssignment;
+  kind: ACTIVITIES.CALL_ASSIGNMENT;
+};
+
+export type SurveyActivity = CampaignActivityBase & {
+  data: ZetkinSurvey;
+  kind: ACTIVITIES.SURVEY;
+};
+
+export type TaskActivity = CampaignActivityBase & {
+  data: ZetkinTask;
+  kind: ACTIVITIES.TASK;
+};
+
+export type EventActivity = CampaignActivityBase & {
+  data: ZetkinEvent;
+  kind: ACTIVITIES.EVENT;
+};
+
+export type CampaignActivity =
+  | CallAssignmentActivity
+  | EventActivity
+  | SurveyActivity
+  | TaskActivity;
+
+export type ActivityOverview = {
+  alsoThisWeek: CampaignActivity[];
+  today: CampaignActivity[];
+  tomorrow: CampaignActivity[];
+};
